Fixed Version (CatTools v806)

+++ script/rm_log.sh
@@ -5,16 +5,16 @@ else
 fi
 for L in $dirlist; do
     if [ ${L} != "Output" ] && [ ${L} != "dataset" ]; then
-#        check=$(./checkFinishedJobs.sh ${L})
-#        if [ "$(echo $check | grep 'All Jobs Finished without Error')" ]; then
+        check=$(./checkFinishedJobs.sh ${L})
+        if [ "$(echo $check | grep 'All Jobs Finished without Error')" ]; then
             echo "$check" 
             rm -rf ${L}/*log*
             rm -rf ${L}/*err*
-#            rm -rf ${L}/*.sh
-#            rm -rf ${L}/job*
+            rm -rf ${L}/*.sh
+            rm -rf ${L}/job*
             rm -rf ${L}/submit*
-#        else
-#            echo "$check" 
-#        fi
+        else
+            echo "$check" 
+        fi
     fi
 done
